feat(pkg) Support creation of Openshift Routes
This WiP work for supporting Openshift Route API which would help users using Openshift to have routes crated when they use the Operator. This work is towards non interference in any other creation of resource, but only if the Openshift Route API exists then create it.

Checks if the Openshift API exists
If it does then creates the Route Resource for the HTTP Service